Implement the Comment model and controller
Let's now generate the `Comment` model and `CommentsController` by running the command: `bin/rails g resource comment post:references body:text`. Note that one of the fields we've added to the comment is `post:references` which adds the foreign key in the `comments` table to associate `comment` records with `post` records.

If we look in the migration file we see the line:
```ruby
t.references :post, index: true, foreign_key: true
```
Notice that there is an `index: true` option added by default. This is because we generally do many queries with the `post_id` foreign key which means indexing will speed up our queries.

We also see that there is an option there `foreign_key: true`. This adds a foreign key constraint in the database. This constraint prevents us from adding a record with invalid `post_id` and prevents us from deleting a `post` record whose `id` is references in some comment.

This is the first part of setting up the association in Rails. When we have a `comment` model created, we can simply get the `post` model by doing: `comment.post`