Set up the association on the Post model
We not set up the association on the `Post` model:
```ruby
has_many :comments, dependent: :destroy
```
This was if we need to fetch all the comments belonging to a single `post` we can do:
```ruby
post.comments
```
Note that we've added a ```ruby dependent: :destroy ``` option in there. This is because in the last step we left the option ```ruby foreign_key: true``` in the migration file for the reference field. This means the database will throw an error if we try to delete a post references by comments. So if by adding this option Rails will delete all the associated comment record first before deleting the `post` record itself using `.destroy` method.

The other option is to use `dependent: :nullify` which will will make the `post_id` in all the associated comment records `null` before deleting the post record.

Note that there is a Rails method called delete that you can call by on a post by doing `post.delete`. Using this method will not delete associated `comment` records even if you have `dependent: :destroy` option in there. I would recommend not using this method to start and stick with using the `destroy` method instead.