Implement the show action for posts
We implement the show action for posts by adding the `show` action in the `PostsController` and adding the `show.html.erb` template in the `views/posts` folder.

We instantiate an `@post` variable that references the `Post` record after finding that record by its `id` using the `find` method.

In the `show.html.erb` template we display the basic attributes of the `@post` which are the `title` and `body`.

We then loop through the comments associated with the post using `@post.comments.each` and then displaying the `body` of every comment in loop block.