Implement the comment creation form
We start by changing our routes by nesting the `resources :comments` within `recources :posts`. This is useful for us because now all the `comments` routes will have `/posts/:post_id` prefix to them.

For instance the comment creation url will not be `/posts/:post_id/comments` and the comment show url will be `/posts/:post_id/comments/:id`

This better follows REST standard as we are creating a comment here for a specific post and not just by itself.

We then instantiate an empty instance variable in the `PostsController` inside the `show` action. We need this as it will help us in creating the form in the `show.html.erb`

We now create the comment form in the `show.html.erb` file. We give the first argument of the form as an array `[@post, @comment]`. This will make the form do the following:

* If ```@comment``` is not persisted to the database the form submit url will be `post_comments_path(@post)` which will be `/posts/:post_id/comments` where `:post_id` is `@post.id`. The method of the form will be `POST`. This means the form will submit to the `create` action in the `CommentsController`

* If ```@comment``` is persisted to the database the form will submit to `post_comment_path(@post, @comment)` which will be `/posts/:post_id/comments/:id` where `:post_id` is `@post.id` and `:id` is `@comment.id`. The method of the form will be `PATCH`. This means that the form will submit to the `update` action in the `CommentsController`.