Implement the comment create action
We start implementing the create action in the `CommentsController` by defining the method and then finding the post record we're creating the comment for. Note that `params[:post_id]` is coming from the URL. The URL that submits to this action is `/posts/:post_id/comments`.

We then instantiate an `@comment` variable with the params submitted taking into account using strong parameters.

We then associate the `@comment` record with the `@post` using `@comment.post = @post`.

If the `@comment` records saves correctly then we redirect back to the `@post` show page which is `post_path(@post)`.

If the `@comment` doesn't save correctly then we render the `show.html.erb` in the `/views/posts` folder. For this we have to use the full path for it because if we do `render :show` it will render `/views/comments/show.html.erb`